feat(opstack): delegate refundGas to OpStackPostSettlementPlan

+++ bcos-evm/opstack/fee/OpStackPostSettlementInputs.h
@@ -19,7 +19,7 @@ inline OpStackPostSettlementInputs toOpStackPostSettlementInputs(
                 view.web3TypedTxKind(), view.hasGasFeeCap()},
             ctx.originalGasLimit),
         .gasUsed = settled.gasUsed,
-        .gasRemaining = settled.gasRemaining,
+        .gasRemaining = static_cast<int64_t>(settled.gasRemaining),
         .blockTime = static_cast<uint64_t>(view.blockInfo().timestamp),
         .l1CostCharged = sidecar.l1CostCharged,
         .operatorCostLimit = sidecar.operatorCostLimit,
